src/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.35
paho-mqtt==2.1.0
pyjwt==2.9.0
python-multipart==0.0.12
